backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
openai>=1.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
httpx>=0.27.0
sse-starlette>=2.0.0
dashscope>=1.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0

